✨ Move fetchRegionRank function to leaderboardUtils and update import in route
